+++ CHANGELOG.rst
@@ -10,6 +10,9 @@
   Improve xitrum.util.SeriDeseri
 * `#474 <https://github.com/xitrum-framework/xitrum/issues/474>`_
   Add isEmpty to SessionVar for convenience
+* `#475 <https://github.com/xitrum-framework/xitrum/issues/475>`_
+  Make writing before filters more convenient:
+  If a filter responds something, the main action will not be called
 * `#462 <https://github.com/xitrum-framework/xitrum/issues/462>`_
   Switch back to Twitter's Chill (from chill-scala) because it now supports Scala 2.11
 * `#461 <https://github.com/xitrum-framework/xitrum/issues/461>`_